Add files via upload
In the E-Commerce shopping experience, it is essential to deliver the product promptly to the customer. 'FastTrack' courier company stores the relevant data of its clients and parcels in the form of a dictionary. Create a dictionary for storing shipment information in key-value pairs. Shipment id is used as a key and list of other attributes like sender, receiver, start date, Delivery Date, Sender_location, Receiver_location, Delivery status, Shipping cost is associated with shipment id.

Skills and Tools

Python, Numpy, Pandas